Added full support for -d (dump) option in MCell
The "-d #" command line parameter is an optional argument specifying a
"dump level" for MCell. The "dump_level" can be any number, and it's up
to any section of code to determine what to do with any particular value.

In this commit, the viz_output code will print additional information as
follows:

  dump_level >= 10:  Prints all new Graph Patterns as they are detected
  dump_level >= 20:  Prints the molcomp_loc_array (with molecule structure)
  dump_level >= 50:  Prints message at top and bottom of viz output loop

If the -d option is not set, it defaults to 0, and all output should be
the same as it was before this change.

+++ src/argparse.c
@@ -161,19 +161,19 @@ int argparse_init(int argc, char *const argv[], struct volume *vol) {
       break;
 
     case 'd': /* -dump */
-      vol->dump_level = (int)strtol(optarg, &endptr, 0);
+      vol->dump_level = strtol(optarg, &endptr, 0);
       if (endptr == optarg || *endptr != '\0') {
         argerror("Dump level must be an integer: %s", optarg);
         return 1;
       }
 
       if (vol->dump_level < 0) {
-        argerror("Dump level %d is less than 0", (int)vol->dump_level);
+        argerror("Dump level %ld is less than 0", vol->dump_level);
         return 1;
       }
 
       if (vol->dump_level > 0) {
-        fprintf(stdout, "Dump level has been set to %d\n", vol->dump_level);
+        fprintf(stdout, "Dump level has been set to %ld\n", vol->dump_level);
       }
       break;
